# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenPeople API",
    "version": "0.1.0",
    "description": "Unofficial OpenAPI surface for critical endpoints. This file is generated and should stay in sync with route schemas."
  },
  "servers": [
    {
      "url": "https://app.openpeople.ai",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "GatewayKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "op_sk_*",
        "description": "Gateway API key header: Authorization: Bearer op_sk_..."
      },
      "ServiceTokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "op_srv_*",
        "description": "Service token header: Authorization: Bearer op_srv_..."
      },
      "UserAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "jwt",
        "description": "User session JWT (via Supabase) for dashboard clients"
      },
      "VaultTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-vault-token",
        "description": "Quick-upload token header (long-lived device token)"
      }
    },
...
    "schemas": {
...
      "Tenant": {
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "suspended"
            ]
          },
          "tier": {
            "type": "string",
            "enum": [
              "free",
              "pro",
              "enterprise"
            ],
            "nullable": true
          },
          "settings": {
            "type": "object",
            "additionalProperties": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TenantCreateRequest": {
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "tier": {
            "type": "string"
          },
          "settings": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "TenantUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tier": {
            "type": "string"
          },
          "settings": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "full_name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "admin",
              "member",
              "viewer",
              "super_admin"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled",
              "invited"
            ]
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          },
          "mfa_enabled": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserInviteRequest": {
        "type": "object",
        "required": [
          "email",
          "role",
          "tenant_id"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "full_name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "UserUpdateRequest": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "VaultFile": {
        "type": "object",
        "required": [
          "id",
          "filename",
          "size_bytes",
          "content_type",
          "vault_id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "vault_id": {
            "type": "string",
            "format": "uuid"
          },
          "folder_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "filename": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer"
          },
          "content_type": {
            "type": "string"
          },
          "ai_summary": {
            "type": "string",
            "nullable": true
          },
          "ai_tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "ai_category": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "active",
              "archived",
              "deleted"
            ]
          }
        }
      },
      "VaultUploadInitRequest": {
        "type": "object",
        "required": [
          "filename",
          "content_type",
          "size_bytes"
        ],
        "properties": {
          "filename": {
            "type": "string"
          },
          "content_type": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer"
          },
          "folder_id": {
            "type": "string",
            "format": "uuid"
          },
          "checksum": {
            "type": "string"
          }
        }
      },
      "SignedUrlResponse": {
        "type": "object",
        "required": [
          "file_id",
          "upload_url"
        ],
        "properties": {
          "file_id": {
            "type": "string",
            "format": "uuid"
          },
          "upload_url": {
            "type": "string",
            "format": "uri"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/tenants": {
      "get": {
        "summary": "List tenants (super-admin/service token)",
        "security": [
          {
            "ServiceTokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Tenant"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create tenant",
        "security": [
          {
            "ServiceTokenAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tenant created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}": {
      "get": {
        "summary": "Get tenant",
        "security": [
          {
            "ServiceTokenAuth": []
          },
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update tenant",
        "security": [
          {
            "ServiceTokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "summary": "List users (tenant scoped)",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Invite user",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInviteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User invited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "patch": {
        "summary": "Update user role/status",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vault/files": {
      "get": {
        "summary": "List vault files",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Files",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VaultFile"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Request signed upload URL",
        "security": [
          {
            "UserAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaultUploadInitRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Signed URL returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedUrlResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}